+++ CHANGELOG.md
@@ -13,9 +13,11 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+
 ### Added
 - PixPullRequest resource
 - PixPullSubscription resource
+- Deprecated create pix infraction
 
 ## [0.20.1] - 2026-02-05
 ### Fixed